Update to Mathematica 12.
The folks at Wolfram are terrible at writing decent Linux GUIs.  The GUI
font size appear to be hardcoded in Mathematica 12.  Thus, if a default
sans-serif font is set in ~/.fonts.conf (or an equivalent), and if that
font has a smaller width than DejaVu Sans, the fonts in the GUI will
appear too small.  One workaround is running Mathematica in a sandboxed
environment (using firejail) and prevent it from accessing ~/.fonts.conf.
But I found it simpler to not set a default sans-serif font in
~/.fonts.conf.  It's not like I'm making use of it anyway.