Vim: better math snippets
Earlier, a "quick snippet" would've have been triggered even if we
weren't inside a math environment.  This was annoying, e.g., a snippet
would be triggered each time I made an equation label of the form
\label{eq:name_name} because of the underscore.

This has now been fixed.  Now we use a vimtex function to check if we're
inside a math environment before triggering snippet insertion.